app/requirements.txt
django==2.2.6
djangorestframework==3.11.0
gunicorn==19.9.0
pytz==2020.1
sqlparse==0.3.1
django-mysql==2.2.0
mysqlclient==1.4.6
numpy==1.18.4
Pillow==7.1.2
requests==2.23.0
djoser==2.0.3 #for authentication